% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonAccess.R
\name{jsonAccess}
\alias{jsonAccess}
\title{Access to the value of a JSON string}
\usage{
jsonAccess(json, path, toJSON = FALSE, fromJSON = FALSE)
}
\arguments{
\item{json}{a JSON string}

\item{path}{character vector, sequence of keys}

\item{toJSON}{logical, whether to apply \code{\link[jsonlite]{toJSON}} to
the input}

\item{fromJSON}{logical, whether to apply \code{\link[jsonlite]{fromJSON}} to
the output}
}
\value{
A json string if \code{fromJSON=FALSE}, a R object if \code{fromJSON=TRUE}.
}
\description{
Access to the value of a JSON string by giving a sequence
of keys.
}
\note{
Instaed of using the option \code{toJSON=TRUE}, you can apply
\code{\link[jsonlite]{toJSON}} before; this allows you to apply it
with its options. Same remark for the option \code{toJSON=FALSE}.
}
\examples{
json <- "{\\"a\\":{\\"b\\":8,\\"c\\":[1,2]},\\"b\\":null}"
jsonAccess(json, "a")
jsonAccess(json, "b")
jsonAccess(json, "X")
jsonAccess(json, c("a","c"))
# get the result as a R object:
jsonAccess(json, c("a","c"), fromJSON=TRUE)
# same as:
jsonAccess(json, c("a","c")) \%>\% jsonlite::fromJSON(.)
}
\seealso{
\code{\link{jsonElemsAt}}
}
