% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonAccess.R
\name{jsonElemAt}
\alias{jsonElemAt}
\title{Extract element of a JSON array}
\usage{
jsonElemAt(json, index, toJSON = FALSE, fromJSON = FALSE)
}
\arguments{
\item{json}{json string}

\item{index}{integer}

\item{toJSON}{logical, whether to apply \code{\link[jsonlite]{toJSON}} to
the input}

\item{fromJSON}{logical, whether to apply \code{\link[jsonlite]{fromJSON}} to
the output}
}
\value{
A JSON string if \code{fromJSON=FALSE}, a R object if \code{fromJSON=TRUE}.
}
\description{
Returns the element of a JSON array by giving the desired
index.
}
\examples{
json <- "[0,1,\\"hello\\",null,\\"bird\\"]"
jsonElemAt(json, 2)
jsonElemAt(json, 3)
jsonElemAt(json, 10) == "null"
jsonElemsAt(json, 1:4) \%>\% jsonElemAt(3)
jsonElemAt("{\\"a\\":{\\"b\\":8}}", 0) == "null"
}
\seealso{
\code{\link{jsonElemsAt}}, \code{\link{jsonElemAt_ifNumber}}, \code{\link{jsonAccess}}
}
