% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonAccess.R
\name{jsonElemAt_ifArray}
\alias{jsonElemAt_ifArray}
\title{Extract an array from a JSON array}
\usage{
jsonElemAt_ifArray(json, index, toJSON = FALSE, fromJSON = FALSE)
}
\arguments{
\item{json}{JSON string}

\item{index}{integer}

\item{toJSON}{logical, whether to apply \code{\link[jsonlite]{toJSON}} to
the input}

\item{fromJSON}{logical, whether to apply \code{\link[jsonlite]{fromJSON}} to
the output}
}
\value{
A JSON string if \code{fromJSON=FALSE}, a R object if \code{fromJSON=TRUE}.
}
\description{
Returns the array in a JSON array by giving its index;
returns \code{"null"} if there is no array at this index.
}
\examples{
json <- "[1,[{},2],null,\\"bird\\"]"
jsonElemAt_ifArray(json, 2)
jsonElemAt_ifArray(json, 1) == "null"
jsonElemAt_ifArray(json, 3)
}
\seealso{
\code{\link{jsonElemAt}}, \code{\link{jsonAccess}}
}
