% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonAccess.R
\name{jsonElemAt_ifNumber}
\alias{jsonElemAt_ifNumber}
\title{Extract a number from a JSON array}
\usage{
jsonElemAt_ifNumber(json, index, toJSON = FALSE, fromJSON = FALSE)
}
\arguments{
\item{json}{JSON string}

\item{index}{integer}

\item{toJSON}{logical, whether to apply \code{\link[jsonlite]{toJSON}} to
the input}

\item{fromJSON}{logical, whether to apply \code{\link[jsonlite]{fromJSON}} to
the output}
}
\value{
A JSON string if \code{fromJSON=FALSE}, a R object if \code{fromJSON=TRUE}.
}
\description{
Returns the number in a JSON array by giving its index;
returns \code{"null"} if there is no number at this index.
}
\examples{
json <- "[1,0,\\"hello\\",null,\\"bird\\"]"
jsonElemAt_ifNumber(json, 0)
jsonElemAt_ifNumber(json, 2) == "null"
jsonElemAt_ifNumber(json, 3)
}
\seealso{
\code{\link{jsonElemAt}}, \code{\link{jsonAccess}}
}
