% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonAccess.R
\name{jsonElemAt_ifString}
\alias{jsonElemAt_ifString}
\title{Extract a string from a JSON array}
\usage{
jsonElemAt_ifString(json, index, toJSON = FALSE, fromJSON = FALSE)
}
\arguments{
\item{json}{JSON string}

\item{index}{integer}

\item{toJSON}{logical, whether to apply \code{\link[jsonlite]{toJSON}} to
the input}

\item{fromJSON}{logical, whether to apply \code{\link[jsonlite]{fromJSON}} to
the output}
}
\value{
A JSON string if \code{fromJSON=FALSE}, a R object if \code{fromJSON=TRUE}.
}
\description{
Returns the string in a JSON array by giving its index;
returns \code{"null"} if there is no string at this index.
}
\examples{
json <- "[1,\\"hello\\",null,\\"bird\\"]"
jsonElemAt_ifString(json, 1)
jsonElemAt_ifString(json, 0) == "null"
}
\seealso{
\code{\link{jsonElemAt}}, \code{\link{jsonAccess}}
}
