% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonAccess.R
\name{jsonElemsAt}
\alias{jsonElemsAt}
\title{Sub-array of a JSON array}
\usage{
jsonElemsAt(json, indices, toJSON = FALSE, fromJSON = FALSE)
}
\arguments{
\item{json}{json string}

\item{indices}{integer vector}

\item{toJSON}{logical, whether to apply \code{\link[jsonlite]{toJSON}} to
the input}

\item{fromJSON}{logical, whether to apply \code{\link[jsonlite]{fromJSON}} to
the output}
}
\value{
A json string if \code{fromJSON=FALSE}, a R object if \code{fromJSON=TRUE}.
}
\description{
Returns a sub-array of a JSON array by giving the desired
indices.
}
\examples{
json <- "{\\"a\\":{\\"b\\":8,\\"c\\":[0,1,\\"hello\\"]},\\"b\\":null}"
( arr <- jsonAccess(json, c("a","c")) )
jsonElemsAt(arr, c(0,2))
jsonElemsAt(arr, 2)
jsonElemsAt(arr, c(1,3))
# chain with the pipe operator:
jsonAccess(json, c("a","c")) \%>\% jsonElemsAt(2)
}
\seealso{
\code{\link{jsonElemAt}}, \code{\link{jsonAccess}}
}
