% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonAccess.R
\name{jsonMembers}
\alias{jsonMembers}
\title{Members of a JSON string}
\usage{
jsonMembers(json, toJSON = FALSE, fromJSON = FALSE)
}
\arguments{
\item{json}{JSON string}

\item{toJSON}{logical, whether to apply \code{\link[jsonlite]{toJSON}} to
the input}

\item{fromJSON}{logical, whether to apply \code{\link[jsonlite]{fromJSON}} to
the output}
}
\value{
A JSON string if \code{fromJSON=FALSE}, a R object if \code{fromJSON=TRUE}.
}
\description{
Returns the members of a JSON string as a JSON string array.
}
\examples{
json <- "{\\"a\\":{\\"b\\":8,\\"c\\":[1,2]},\\"b\\":[10,11]}"
jsonMembers(json)
}
\seealso{
\code{\link{jsonElemAt}}, \code{\link{jsonAccess}}
}
