% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonAccess.R
\name{jsonPretty}
\alias{jsonPretty}
\title{Pretty JSON string}
\usage{
jsonPretty(json, indent = "indent2", numformat = "generic",
  toJSON = FALSE)
}
\arguments{
\item{json}{JSON string}

\item{indent}{indentation per level of nesting; choices are
\code{"indent0"} (remove \strong{all} white spaces), \code{"indent1"},
\code{"indent2"}, \code{"indent3"}, \code{"indent3"}, and
\code{"tab"} (tabulation)}

\item{numformat}{number format; choices are:
\itemize{
     \item \code{"generic"}: integer literals for integers (1, 2, ...),
     simple decimals for fractional values between 0.1 and 9999999,
     and scientific notation otherwise
     \item \code{"decimal"}: standard decimal notation
     \item \code{"scientific"}: scientific notation
}}

\item{toJSON}{logical, whether to apply \code{\link[jsonlite]{toJSON}} to
the input}
}
\value{
A JSON string.
}
\description{
Returns a JSON string in pretty format.
}
\examples{
json <- "{\\"a\\": {\\"b\\": 8, \\"c\\": [1,2]}, \\"b\\": [10.14, 11.618]}"
jsonPretty(json) \%>\% cat
jsonPretty(json, indent="indent0") \%>\% cat
jsonPretty(json, indent="tab", numformat="scientific") \%>\% cat
}
